Added a bunch of `raise_...` methods to improve readability.
Since, historically, the code would raise different error messages depending of if it was dealing with long or short style options, and since I didn't change this to maintain 100% compatibility, this was the last obstacle to readability.
The newly added methods decide from their parameters, which error message to trigger.